requirements.txt
bleach==3.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
httpie==2.0.0
idna==2.9
Markdown==3.1.1
Pillow==6.0.0
psycopg2==2.8.2
Pygments==2.6.1
python-decouple==3.1
pytz==2018.7
requests==2.23.0
six==1.12.0
static3==0.7.0
Unipath==1.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==4.1.2
